AI_Syllabus_Planner_final/requirements.txt
flask
Flask-SQLAlchemy
Flask-Login
werkzeug
python-dotenv
google-generativeai
